Rename arrow module to stochastic_arrow
