+++ stochastic_arrow/analysis/plotting.py
