+++ stochastic_arrow/arrow.py
