+++ stochastic_arrow/arrowhead.pyx
