+++ stochastic_arrow/math.py
