+++ stochastic_arrow/mersenne.c
