+++ stochastic_arrow/mersenne.h
