+++ stochastic_arrow/obsidian.h
