+++ stochastic_arrow/reference.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function
 import numpy as np
 from six import moves
 
-from arrow.math import multichoose
+from stochastic_arrow.math import multichoose
 
 
 def derive_reactants(stoichiometric_matrix):
