+++ stochastic_arrow/test/test_hang.py
@@ -26,7 +26,7 @@
 
 import os
 
-from arrow import StochasticSystem
+from stochastic_arrow import StochasticSystem
 import numpy as np
